backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
PyJWT==2.10.1
pydantic==2.11.7
